% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cov2}
\alias{cov2}
\title{Calculo da covariancia normalizada por 1/n}
\usage{
cov2(x, na.rm = FALSE)
}
\arguments{
\item{x}{matriz numerica na qual se deseja calcular a covariancia}

\item{na.rm}{logico indicando se valores NA devem ser removidos antes do calculo}
}
\value{
covariancia calculada
}
\description{
Funcao alternativa para calcular covariancia atraves do estimador "n"
}
