% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dummyseries}
\alias{dummyseries}
\title{Series Exemplo para Modelos PAR}
\format{
\subsection{dummyseries}{

Uma serie temporal multivariada (mts) com 3 colunas e 1092 observacoes, iniciando em Janeiro
de 1931 com frequencia 12 meses
}
}
\usage{
dummyseries
}
\description{
Series de vazoes incrementais mensais de Camargos, Furnas e Igarapava
}
\keyword{datasets}
