% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{filter_series}
\alias{filter_series}
\title{Filtra Serie Temporal Com Modelo PAR}
\usage{
filter_series(object)
}
\arguments{
\item{object}{objeto da classe \code{par} contendo o modelo ajustado}
}
\value{
serie temporal filtrada
}
\description{
Filtra uma serie temporal utilizando um modelo PAR ajustado
}
