% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-par_a.r
\name{fit_par_a}
\alias{fit_par_a}
\alias{fit_par_a_yulewalker}
\title{Ajuste Por Estacao De \code{par_a}}
\usage{
fit_par_a(serie, serie_A, m, p, metodo, ...)

fit_par_a_yulewalker(serie, serie_A, m, p, max_p, ...)
}
\arguments{
\item{serie}{serie temporal sazonal (bruta)}

\item{serie_A}{componente anual ja calculada, repassada para evitar recomputo a cada estacao}

\item{m}{estacao do ano a ajustar}

\item{p}{ordem autorregressiva, ou \code{"auto"} para identificacao automatica}

\item{metodo}{metodo de estimacao}

\item{...}{demais argumentos repassados ao metodo}

\item{max_p}{ordem maxima na identificacao automatica}
}
\value{
lista com \code{phis} (coeficientes autorregressivos) e \code{psi} (coeficiente anual)
}
\description{
Funcoes internas para ajuste do modelo periodico anual de uma unica estacao
}
