% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-par_a.r
\name{idordem_cacf}
\alias{idordem_cacf}
\title{Identificacao Automatica De Ordem De Modelos \code{par_a}}
\usage{
idordem_cacf(serie, serie_A, m, max_p)
}
\arguments{
\item{serie}{serie temporal sazonal (bruta)}

\item{serie_A}{componente anual ja calculada}

\item{m}{estacao do ano}

\item{max_p}{numero maximo de lags a considerar}
}
\value{
ordem identificada
}
\description{
Funcao interna que identifica a ordem autorregressiva via FACP condicional (\code{percacf}), seguindo o
mesmo procedimento de \code{\link[=idordem_yulewalker]{idordem_yulewalker()}}: mantem os lags cujo coeficiente excede a banda de 95\%
(\verb{+/- z_0.975 / sqrt(n)}), ate \code{max_p}.
}
