% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-par.r
\name{idordem_yulewalker}
\alias{idordem_yulewalker}
\title{Identificacao Automatica De Ordem De Modelos Autorregressivos Periodicos}
\usage{
idordem_yulewalker(serie, m, max_p)
}
\arguments{
\item{serie}{serie temporal sazonal}

\item{m}{estacao do ano para a qual identificar a ordem}

\item{max_p}{numero maximo de lags a considerar}
}
\value{
ordem identificada
}
\description{
Funcao interna para identificar automaticamente a ordem de modelos autorregressivos periodicos
}
