% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{pad_series}
\alias{pad_series}
\title{Padding De Serie Temporal}
\usage{
pad_series(serie, pad = NA)
}
\arguments{
\item{serie}{serie temporal a ser preenchida}

\item{pad}{valor a ser utilizado para preencher as pontas (default: NA)}
}
\value{
serie temporal preenchida com NA nas pontas
}
\description{
Preenche uma serie temporal com NA no inicio e no fim para completar ciclos sazonais
}
