% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-par.r
\name{par}
\alias{par}
\title{Estimacao De \code{par}s}
\usage{
par(serie, ps = "auto", max_ps = frequency(serie) - 1, metodo = "YuleWalker")
}
\arguments{
\item{serie}{serie temporal com sazonalidade}

\item{ps}{ordens de cada modelo periodico. Veja Detalhes}

\item{max_ps}{se nao forem fornecidas ordens, valor maximo a ser considerado para identificacao
automatica}

\item{metodo}{metodo de estimacao. Atualmente, apenas "Yule-Walker" é suportado}
}
\value{
objeto da classe \code{par}, uma lista contendo os seguintes elementos:
* \code{phis}: lista contendo os coeficientes de cada modelo autorregressivo periodico
* \code{sigma2}: vetor de comprimento igual a frequencia sazonal com a variancia (escala bruta) da
inovacao de cada estacao, calculada em forma fechada via Yule-Walker
* \code{x}: serie temporal utilizada na estimacao
* \code{call}: chamada da funcao
}
\description{
Estimacao de modelos autorregressivos periodicos
}
\details{
O argumento \code{ps} pode ser ou um escalar ou um vetor. No caso de ser escalar, sera repetido para N
vezes, onde N é a frequencia sazonal da serie. Caso seja um vetor, deve ter comprimento igual a
frequencia sazonal da serie. Em ambos os casos, o valor em cada posicao indica a ordem do modelo
autorregressivo periodico para a estacao correspondente. Sendo escalar ou vetor, o valor "auto"
pode ser utilizado para indicar que a ordem deve ser identificada automaticamente utilizando o
criterio correspondente a \code{metodo}: se Yule-Walker, analise das pacfs periodicas.
}
