% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-par_a.r
\name{par_a}
\alias{par_a}
\title{Estimacao De \code{par_a}s}
\usage{
par_a(serie, ps = "auto", max_ps = frequency(serie) - 1, metodo = "YuleWalker")
}
\arguments{
\item{serie}{serie temporal com sazonalidade}

\item{ps}{ordens autorregressivas de cada modelo periodico. Veja \code{\link[=par]{par()}}}

\item{max_ps}{valor maximo a considerar na identificacao automatica de ordem}

\item{metodo}{metodo de estimacao. Atualmente, apenas "Yule-Walker" e suportado}
}
\value{
objeto de classe \code{c("par_a", "par", "list")}, um \code{par} cujos \code{phis} ja incorporam o termo
anual desdobrado, acrescido de:
* \code{psi}: vetor com o coeficiente anual estimado de cada estacao
* \code{ps}: vetor com a ordem autorregressiva identificada de cada estacao
* \code{annual}: serie temporal das medias anuais utilizadas no ajuste (para validacao)
}
\description{
Estimacao de modelos autorregressivos periodicos com componente anual (PAR(p)-A)
}
\details{
Um \code{par_a} e ajustado resolvendo, por estacao, o sistema de Yule-Walker \emph{aumentado} pela componente
anual \code{A} (media dos \code{frequency(serie)} valores imediatamente anteriores). O coeficiente anual \code{psi}
resultante e entao "desdobrado" nos coeficientes de cada um dos 12 lags do ultimo ano, de modo que o
objeto retornado e, mecanicamente, um \code{par} de ordem \code{frequency(serie)}: todos os metodos de \code{par}
(\code{fitted}/\code{residuals}/\code{predict}/\code{simulate}) funcionam por heranca, sem recursao especializada.

O argumento \code{ps} segue a mesma convencao de \code{\link[=par]{par()}}: escalar (replicado) ou vetor de comprimento
igual a frequencia sazonal; \code{"auto"} identifica a ordem automaticamente.
}
\examples{
serie <- ts(dummyseries[, 2], start = c(1931, 1), frequency = 12)
mod <- par_a(serie)

}
