% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-par.r
\name{fitted.par}
\alias{fitted.par}
\alias{fitted.values.par}
\alias{residuals.par}
\alias{predict.par}
\alias{simulate.par}
\title{Metodos De Modelos \code{par}}
\usage{
\method{fitted}{par}(object, ...)

\method{fitted}{values.par}(object, ...)

\method{residuals}{par}(object, ...)

\method{predict}{par}(object, n.ahead, ...)

\method{simulate}{par}(object, nsim, seed = 1234, n.ahead = 1, ...)
}
\arguments{
\item{object}{objeto da classe \code{par} contendo o modelo ajustado}

\item{...}{argumentos adicionais}

\item{n.ahead}{numero de passos a frente a serem previstos}

\item{nsim}{inteiro, numero de simulacoes a realizar}

\item{seed}{semente para simulacao}
}
\value{
varios, dependendo do metodo
}
\description{
Metodos S3 para objetos da classe \code{par}
}
