% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parmodels-package.r
\docType{package}
\name{parmodels-package}
\alias{parmodels}
\alias{parmodels-package}
\title{parmodels: Conjunto de funcionalidades para estimacao e uso de modelos autorregressivos periodicos}
\description{
Este pacote contem toda a funcionalidade necessaria para estimacao, incluindo selecao de ordem automatica, e uso de modelos autorregressivos periodicos (PAR). Funcoes para computo e visualizacao das autocorrelacoes (comuns e parciais) periodicas (PACF) sao tambem disponibilizadas. Modelos do tipo PAR-A, com componente anual, tambem sao suportados, bem como as funcoes de autocorrelacao condicionais envolvidas na sua estimacao.
}
\author{
\strong{Maintainer}: Lucas Khenayfis \email{lucaskhenayfis@gmail.com}

}
\keyword{internal}
