% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfs.r
\name{build_RHO_A}
\alias{build_RHO_A}
\alias{build_rho_A}
\title{Funcoes Auxiliares De \code{percacf}}
\usage{
build_RHO_A(serie, serie_anual, m, lag_max, est)

build_rho_A(serie, serie_anual, m, lag_max, est)
}
\arguments{
\item{serie}{serie temporal sazonal (ja padronizada)}

\item{serie_anual}{componente anual (ja padronizada) alinhada a \code{serie}}

\item{m}{estacao do ano para a qual calcular}

\item{lag_max}{numero maximo de lags a considerar}

\item{est}{um de \code{c("n", "n-1")} indicando qual denominador utilizar para desvio padrao}
}
\description{
Funcoes internas para montagem do sistema de Yule-Walker aumentado (com componente anual)
}
\details{
Reaproveitam \code{build_RHO}/\code{build_rho} para o bloco Z-Z e apenas acrescentam a borda anual: a
componente anual entra como ultima linha/coluna do sistema (layout CEPEL / \code{methodology_report} 4).
}
