% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfs.r
\name{perpacf}
\alias{perpacf}
\title{PACF Periodica De Series Temporais}
\usage{
perpacf(
  serie,
  m,
  lag_max = frequency(serie) - 1,
  est = c("n", "n-1"),
  plot = FALSE
)
}
\arguments{
\item{serie}{serie temporal sazonal. Sera padronizada internamente}

\item{m}{estacao do ano para a qual calcular a PACF}

\item{lag_max}{numero maximo de lags a considerar. Default é frequencia da serie - 1}

\item{est}{um de \code{c("n", "n-1")} indicando qual denominador utilizar para desvio padrao}

\item{plot}{se TRUE, plota a PACF periodica}
}
\value{
objeto da classe \code{periodic_pacf}, uma lista contendo os seguintes elementos:
* \code{phi}: vetor com os valores da PACF periodica
* \code{n.used}: numero de observacoes utilizadas no calculo
* \code{m}: estacao do ano para a qual a PACF foi calculada
* \code{lag_max}: numero maximo de lags considerados
* \code{rho}: vetor com os valores da funcao de autocorrelacao (ACF) periodica
* \code{RHO}: matriz com os valores da matriz de autocorrelacao (ACF) periodica
}
\description{
Calcula a funcao de autocorrelacao parcial periodica (PACF) de uma serie temporal sazonal
}
