% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfs.r
\name{build_RHO}
\alias{build_RHO}
\alias{build_rho}
\title{Funcoes Auxiliares De \code{perpacf}}
\usage{
build_RHO(serie, m, lag_max, est)

build_rho(serie, m, lag_max, est)
}
\arguments{
\item{serie}{serie temporal sazonal}

\item{m}{estacao do ano para a qual calcular a PACF}

\item{lag_max}{numero maximo de lags a considerar. Default é frequencia da serie - 1}

\item{est}{um de \code{c("n", "n-1")} indicando qual denominador utilizar para desvio padrao}
}
\description{
Funcoes internas para calculo da funcao de autocorrelacao periodica
}
\details{
Estas funcoes calculam o vetor \code{rho} e a matriz \code{RHO} necessarios para execucao do metodo de
Yule-Walker para calculo de autocorrelacoes parciais periodicas.
}
