% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.periodic_pacf}
\alias{plot.periodic_pacf}
\title{Plota a PACF periodica}
\usage{
\method{plot}{periodic_pacf}(x, ...)
}
\arguments{
\item{x}{objeto da classe \code{periodic_pacf}}

\item{...}{argumentos adicionais para a funcao \code{plot}}
}
\value{
nenhum, apenas plota o grafico
}
\description{
Plota a funcao de autocorrelacao parcial periodica calculada pela funcao \code{perpacf}.
}
