% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot_simulation}
\alias{plot_simulation}
\title{Plota Simulacao De Modelo \code{par}}
\usage{
plot_simulation(simul, predicted = NULL, ref = NULL)
}
\arguments{
\item{simul}{saida de uma chamada de \code{simulate}}

\item{predicted}{opcional, serie temporal de previsao a ser plotada como referencia}

\item{ref}{opcional, serie temporal de observado a ser plotada como referencia}
}
\value{
nao ha retorno, apenas print do plot R base
}
\description{
Plota grafico combinando todas as simulacoes em \code{simul}, com verificado opcional por cima
}
