% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{predict_series}
\alias{predict_series}
\title{Previsao Com Modelo PAR}
\usage{
predict_series(object, n.ahead)
}
\arguments{
\item{object}{objeto da classe \code{par} contendo o modelo ajustado}

\item{n.ahead}{numero de periodos a serem previstos}
}
\value{
serie temporal contendo as previsoes
}
\description{
Realiza previsao de uma serie temporal utilizando um modelo PAR ajustado
}
