% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{scale_by_season}
\alias{scale_by_season}
\title{Normaliza Serie Sazonalmente}
\usage{
scale_by_season(serie, est = "n", means = NULL, sds = NULL)
}
\arguments{
\item{serie}{serie temporal com sazonalidade para normalizar}

\item{est}{um de \code{c("n", "n-1")} indicando qual denominador utilizar para desvio padrao}

\item{means, sds}{medias e desvios padrao sazonais ja calculadas (opcional)}
}
\value{
lista de dois elementos: serie temporal padronizada sazonalmente e lista contendo medias
e desvios padrao sazonais
}
\description{
Normaliza uma serie temporal com sazonalidade utilizando medias e desvios padrao por estacao
}
\details{
Caso \code{means} e \code{sds} sejam fornecidos, devem ser vetores com comprimento igual a a frequencia
sazonal da serie, em que a primeira posicao corresponde à primeira estacao e assim por diante.
}
