% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{sd2}
\alias{sd2}
\title{Calculo do desvio padrao normalizado por 1/n}
\usage{
sd2(x, na.rm = FALSE)
}
\arguments{
\item{x}{vetor numerico do qual se deseja calcular o desvio padrao}

\item{na.rm}{logico indicando se valores NA devem ser removidos antes do calculo}
}
\value{
desvio padrao calculado
}
\description{
Funcao alternativa para calcular desvio padrao atraves do estimador "n"
}
