% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{ts2matrix}
\alias{ts2matrix}
\title{Converte Serie Temporal em Matriz}
\usage{
ts2matrix(serie)
}
\arguments{
\item{serie}{serie temporal a ser convertida}
}
\value{
matriz com as observacoes organizadas por linhas de ciclos sazonais
}
\description{
Converte uma serie temporal em matriz, preenchendo com NA nas pontas para completar ciclos sazon
}
