% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{calcula_medias_anuais}
\alias{calcula_medias_anuais}
\title{Calcula Medias Anuais}
\usage{
calcula_medias_anuais(serie)
}
\description{
Retorna \code{ts} das medias relativas ao ciclo sazonal completo antecedente a cada observacao
}
