% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-par.r
\name{idordem_yulewalker}
\alias{idordem_yulewalker}
\title{Identificacao Automatica De Ordem De Modelos Autorregressivos Periodicos}
\usage{
idordem_yulewalker(serie, m, max_p)
}
\description{
Funcao interna para identificar automaticamente a ordem de modelos autorregressivos periodicos
}
