% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-par.r
\name{fitted.par}
\alias{fitted.par}
\alias{fitted.values.par}
\alias{residuals.par}
\alias{predict.par}
\title{Metodos De Modelos \code{par}}
\usage{
\method{fitted}{par}(object, ...)

\method{fitted}{values.par}(object, ...)

\method{residuals}{par}(object, ...)

\method{predict}{par}(object, n.ahead, ...)
}
\arguments{
\item{object}{objeto da classe \code{par} contendo o modelo ajustado}

\item{...}{argumentos adicionais}

\item{n.ahead}{numero de passos a frente a serem previstos}
}
\value{
varios, dependendo do metodo
}
\description{
Metodos S3 para objetos da classe \code{par}
}
