% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{run_model_recursion}
\alias{run_model_recursion}
\title{Previsao Com Modelo PAR}
\usage{
run_model_recursion(object, n.ahead, mode = c("prediction", "simulation"))
}
\arguments{
\item{object}{objeto da classe \code{par} contendo o modelo ajustado}

\item{n.ahead}{numero de periodos a serem previstos}
}
\value{
serie temporal contendo as previsoes
}
\description{
Realiza previsao de uma serie temporal utilizando um modelo PAR ajustado
}
